stora vinster

+++ src/main/java/com/example/SecondHandShop/ProductRepo.java
@@ -1,7 +1,13 @@
 package com.example.SecondHandShop;
 
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepo extends CrudRepository <Product, Integer>{
+import java.util.ArrayList;
+import java.util.List;
+
+public interface ProductRepo extends CrudRepository <Product, Long>{
+
+
 
 }
